backend/requirements.txt
langchain
langchain-community==0.2.1
langchain-core==0.2.17
yfinance==0.2.40
quantstats==0.0.62
diskcache==5.6.3
plotly==5.22.0
chainlit
websockets
httpx
requests
ollama
fastapi
uvicorn
cachetools
IPython
starlette
matplotlib
psycopg2-binary
langchain-together==0.1.4
python-dotenv